graphql refactor
